refactor(puffin): reorganize `trace_helper` and `test_utils`
