+++ puffin/src/lib.rs
@@ -17,6 +17,7 @@ pub mod put_registry;
 pub mod stream;
 pub mod test_utils;
 pub mod trace;
+pub mod trace_helper;
 pub mod variable_data;
 
 pub use {libafl, libafl_bolts};
