+++ puffin/src/test_utils.rs
@@ -1,4 +1,5 @@
 use crate::algebra::{Matcher, Term};
+use crate::execution::ExecutionStatus;
 use crate::graphviz::write_graphviz;
 use crate::trace::{Action, Trace};
 
@@ -45,5 +46,21 @@ impl<M: Matcher> Term<M> {
             }
         }
         found
+    }
+}
+
+pub trait AssertExecution {
+    fn expect_crash(self);
+}
+
+impl AssertExecution for Result<ExecutionStatus, String> {
+    fn expect_crash(self) {
+        use ExecutionStatus as S;
+        match self {
+            Ok(S::Failure(_)) | Ok(S::Crashed) => (),
+            Ok(S::Timeout) => panic!("trace execution timed out"),
+            Ok(S::Success) => panic!("expected trace execution to crash, but succeeded"),
+            Err(reason) => panic!("trace execution error: {reason}"),
+        }
     }
 }
