+++ tlspuffin/src/tls/seeds.rs
@@ -1754,7 +1754,7 @@ macro_rules! corpus {
     ( $( $func:ident : cfg($( $meta:meta),* ) ),* ) => {
         {
             #[cfg(any( $( $( $meta ),* ),* ))]
-            use crate::tls::trace_helper::TraceHelper;
+            use puffin::trace_helper::TraceHelper;
             #[cfg(any( $( $( $meta ),* ),* ))]
             let mut corpus = vec![];
 
@@ -1794,12 +1794,9 @@ pub fn create_corpus() -> Vec<(Trace<TlsQueryMatcher>, &'static str)> {
 
 #[cfg(test)]
 pub mod tests {
-    use puffin::agent::AgentName;
-    use puffin::trace::Action;
-
     use super::*;
-    use crate::put_registry::tls_registry;
-    use crate::tls::trace_helper::TraceHelper;
+    #[allow(unused_imports)]
+    use crate::{test_utils::prelude::*, tls::seeds::*};
 
     #[test_log::test]
     fn test_version() {
@@ -1814,8 +1811,6 @@ pub mod tests {
     #[test_log::test]
     #[cfg(feature = "tls12")]
     fn test_seed_client_attacker12() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_client_attacker12.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1824,8 +1819,6 @@ pub mod tests {
     #[cfg(feature = "transcript-extraction")] // this depends on extracted transcripts -> claims are required
     #[test_log::test]
     fn test_seed_client_attacker() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_client_attacker.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1835,8 +1828,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_client_attacker_auth() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_client_attacker_auth.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1844,8 +1835,6 @@ pub mod tests {
     #[cfg(feature = "tls13")] // require version which supports TLS 1.3
     #[test_log::test]
     fn test_seed_client_attacker_full() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_client_attacker_full.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1854,8 +1843,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_server_attacker_full() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_server_attacker_full.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1865,8 +1852,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_session_resumption_dhe() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_session_resumption_dhe.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1876,8 +1861,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_session_resumption_dhe_full() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_session_resumption_dhe_full.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1887,8 +1870,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_session_resumption_ke() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_session_resumption_ke.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1897,8 +1878,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_successful() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_successful.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1907,8 +1886,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_successful_client_auth() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_successful_client_auth.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1921,8 +1898,6 @@ pub mod tests {
     // mac"  // in case MITM attack did fail
     #[should_panic]
     fn test_seed_successful_mitm() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_successful_mitm.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1931,8 +1906,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_successful_with_ccs() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_successful_with_ccs.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1943,8 +1916,6 @@ pub mod tests {
     #[cfg(not(feature = "boringssl-binding"))]
     #[test_log::test]
     fn test_seed_successful_with_tickets() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         let ctx = seed_successful_with_tickets.execute_trace();
         assert!(ctx.agents_successful());
     }
@@ -1953,8 +1924,6 @@ pub mod tests {
     #[cfg(feature = "tls12")]
     #[cfg(not(feature = "boringssl-binding"))]
     fn test_seed_successful12() {
-        use crate::tls::trace_helper::TraceExecutor;
-
         #[cfg(feature = "tls12-session-resumption")]
         let ctx = seed_successful12_with_tickets.execute_trace();
         #[cfg(not(feature = "tls12-session-resumption"))]
@@ -1964,7 +1933,9 @@ pub mod tests {
 
     #[test_log::test]
     fn test_corpus_file_sizes() {
-        let client = AgentName::first();
+        use puffin::trace::Action;
+
+        let client = puffin::agent::AgentName::first();
         let _server = client.next();
 
         for (trace, name) in create_corpus() {
@@ -1989,7 +1960,9 @@ pub mod tests {
 
     #[test_log::test]
     fn test_term_sizes() {
-        let client = AgentName::first();
+        use puffin::trace::Action;
+
+        let client = puffin::agent::AgentName::first();
         let _server = client.next();
 
         for (name, trace) in [
@@ -2032,8 +2005,8 @@ pub mod tests {
         use puffin::algebra::{set_deserialize_signature, Matcher};
         use puffin::trace::Trace;
 
+        use crate::test_utils::prelude::*;
         use crate::tls::seeds::*;
-        use crate::tls::trace_helper::TraceHelper;
         use crate::tls::TLS_SIGNATURE;
 
         fn test_postcard_serialization<M: Matcher>(trace: Trace<M>) {
